Add files via upload
In Week 1 of our Smart Irrigation System project, we start with the intro of this internship and goes with the basic introduction of machine learning, regression and classification.
We learned how to use Jupyter Notebook to write and run Python code. We also understood how to create and read CSV files which store sensor data like soil moisture and temperature.